requirements.txt
python-dotenv>=1.0.0,<2.0.0

